requirements.txt
transformers>=4.44.0  # >=4.49.0 for SmolVLM2 (train_smolvlm.py)
accelerate>=0.33.0
deepspeed>=0.14.0
peft>=0.12.0
bitsandbytes>=0.43.0
pillow
opencv-python
torch>=2.4.0
torchvision
datasets
sentencepiece
protobuf
matplotlib
qwen-vl-utils
num2words
psutil
